added use strict
broke some of the functions that used the this keyword and became undefined because of it.

Those functions happened to also be arrow functions, so they might unbreak if I used the normal function calling method.

Use strict also showed me that everything needs to at least have the let keyword in front of everything that is a variable and not a parameter

for(e of x) needs to be for(let e of x)
and for i =0 needs to be for let i=0